Collect URL generators in betty.url_generators
