--- betty/tests/project/test_url.py
@@ -1,399 +1,0 @@
-from collections.abc import Mapping, Sequence
-from typing import Any, ClassVar
-
-import pytest
-from babel import Locale
-
-from betty.entity import EntityDefinition
-from betty.entity.collection.pool import EntityPool
-from betty.locale import ResolvableLocale, default_locale, default_locale_tag
-from betty.media_type import MediaType
-from betty.media_types.html import HTML
-from betty.media_types.json import JSON
-from betty.project import Project
-from betty.project.url import (
-    _EntityUrlGenerator,
-    _EntityUrlUrlGenerator,
-    _LocalizedPathUrlUrlGenerator,
-    _StaticPathUrlUrlGenerator,
-    new_project_url_generator,
-)
-from betty.test_utils.conftest import IsolatedProjectFactory
-from betty.test_utils.entity import DummyEntityOne
-
-
-class Test_EntityUrlUrlGenerator:
-    @pytest.mark.parametrize(
-        ("expected", "resource"),
-        [
-            (False, ""),
-            (False, "betty-entity"),
-            (False, "betty-entity://"),
-            (False, "betty-entity://["),
-            (False, f"betty-entity://{DummyEntityOne.plugin().id}"),
-            (False, f"betty-entity://{DummyEntityOne.plugin().id}/"),
-            (True, f"betty-entity://{DummyEntityOne.plugin().id}/my-first-entity"),
-            (False, "/"),
-        ],
-    )
-    async def test_supports(self, expected: bool, resource: Any) -> None:
-        ancestry = EntityPool()
-        sut = _EntityUrlUrlGenerator(
-            ancestry,
-            _EntityUrlGenerator(
-                "https://example.com",
-                "/",
-                {
-                    default_locale: default_locale_tag,
-                },
-                True,
-            ),
-        )
-        assert sut.supports(resource) == expected
-
-    async def test_generate(self) -> None:
-        fragment = "my-first-fragment"
-        locale = "nl-NL"
-        query = {"my_first_query": "my first value"}
-        entity = DummyEntityOne()
-        ancestry = EntityPool()
-        ancestry.add(entity)
-        sut = _EntityUrlUrlGenerator(
-            ancestry,
-            _EntityUrlGenerator(
-                "https://example.com",
-                "/",
-                {
-                    default_locale: default_locale_tag,
-                },
-                True,
-            ),
-        )
-        assert (
-            sut.generate(
-                f"betty-entity://{DummyEntityOne.plugin().id}/{entity.id}",
-                absolute=True,
-                fragment=fragment,
-                locale=locale,
-                media_type=HTML,
-                query=query,
-            )
-            == f"https://example.com/dummy-one/{entity.id}?my_first_query=my+first+value#my-first-fragment"
-        )
-
-
-class Test_LocalizedPathUrlUrlGenerator:
-    @pytest.mark.parametrize(
-        ("expected", "resource"),
-        [
-            (False, ""),
-            (False, "betty"),
-            (False, "betty://"),
-            (False, "betty://["),
-            (True, "betty://without-leading-slash/index.html"),
-            (True, "betty:///with-leading-slash/index.html"),
-            (False, "betty-other://without-leading-slash/index.html"),
-            (False, "betty-other:///with-leading-slash/index.html"),
-            (False, "/"),
-        ],
-    )
-    async def test_supports(
-        self, expected: bool, resource: Any, isolated_project: Project
-    ) -> None:
-        sut = await _LocalizedPathUrlUrlGenerator.new(isolated_project)
-        assert sut.supports(resource) == expected
-
-    _GENERATE_RESOURCES: ClassVar[Sequence[str]] = [
-        "betty://some/path/index.html",
-        "betty:///some/path/index.html",
-    ]
-
-    @pytest.mark.parametrize(
-        (
-            "expected",
-            "resource",
-            "media_type",
-            "absolute",
-            "locale",
-            "additional_project_locale",
-        ),
-        [
-            *[
-                (
-                    "/some/path/index.html",
-                    resource,
-                    media_type,
-                    False,
-                    None,
-                    None,
-                )
-                for resource in _GENERATE_RESOURCES
-                for media_type in [HTML, JSON]
-            ],
-            *[
-                (
-                    "https://example.com/some/path/index.html",
-                    resource,
-                    media_type,
-                    True,
-                    None,
-                    None,
-                )
-                for resource in _GENERATE_RESOURCES
-                for media_type in [HTML, JSON]
-            ],
-            *[
-                (
-                    f"/{default_locale_tag}/some/path/index.html",
-                    resource,
-                    media_type,
-                    False,
-                    None,
-                    "nl-NL",
-                )
-                for resource in _GENERATE_RESOURCES
-                for media_type in [HTML, JSON]
-            ],
-            *[
-                (
-                    "/nl-NL/some/path/index.html",
-                    resource,
-                    media_type,
-                    False,
-                    "nl-NL",
-                    "nl-NL",
-                )
-                for resource in _GENERATE_RESOURCES
-                for media_type in [HTML, JSON]
-            ],
-        ],
-    )
-    async def test_generate(
-        self,
-        expected: str,
-        resource: str,
-        media_type: MediaType,
-        absolute: bool,
-        locale: ResolvableLocale | None,
-        additional_project_locale: Locale | None,
-        isolated_project_factory: IsolatedProjectFactory,
-    ) -> None:
-        locales = [default_locale]
-        if additional_project_locale:
-            locales.append(additional_project_locale)
-        async with isolated_project_factory(locales=locales) as project:
-            sut = await _LocalizedPathUrlUrlGenerator.new(project)
-            assert (
-                sut.generate(
-                    resource,
-                    media_type=media_type,
-                    absolute=absolute,
-                    locale=locale,
-                )
-                == expected
-            )
-
-
-class Test_StaticPathUrlUrlGenerator:
-    _GENERATE_RESOURCES: ClassVar[Sequence[str]] = [
-        "betty-static://some/path/index.html",
-        "betty-static:///some/path/index.html",
-    ]
-
-    @pytest.mark.parametrize(
-        ("expected", "resource"),
-        [
-            (False, ""),
-            (False, "betty-static"),
-            (False, "betty-static://"),
-            (False, "betty-static://["),
-            (True, "betty-static://without-leading-slash/index.html"),
-            (True, "betty-static:///with-leading-slash/index.html"),
-            (False, "betty-static-other://without-leading-slash/index.html"),
-            (False, "betty-static-other:///with-leading-slash/index.html"),
-            (False, "/"),
-        ],
-    )
-    async def test_supports(
-        self, expected: bool, resource: Any, isolated_project: Project
-    ) -> None:
-        sut = await _StaticPathUrlUrlGenerator.new(isolated_project)
-        assert sut.supports(resource) == expected
-
-    @pytest.mark.parametrize(
-        (
-            "expected",
-            "resource",
-            "media_type",
-            "absolute",
-            "locale",
-            "additional_project_locale",
-            "fragment",
-            "query",
-        ),
-        [
-            *[
-                (
-                    "https://example.com/some/path/index.html?my_first_query=my+first+value#my-first-fragment",
-                    resource,
-                    media_type,
-                    True,
-                    locale,
-                    additional_project_locale,
-                    "my-first-fragment",
-                    {"my_first_query": "my first value"},
-                )
-                for resource in _GENERATE_RESOURCES
-                for media_type in [HTML, JSON]
-                for locale in [None, "nl-NL"]
-                for additional_project_locale in [None, "nl-NL"]
-            ],
-            *[
-                (
-                    "/some/path/index.html?my_first_query=my+first+value#my-first-fragment",
-                    resource,
-                    media_type,
-                    False,
-                    locale,
-                    additional_project_locale,
-                    "my-first-fragment",
-                    {"my_first_query": "my first value"},
-                )
-                for resource in _GENERATE_RESOURCES
-                for media_type in [HTML, JSON]
-                for locale in [None, "nl-NL"]
-                for additional_project_locale in [None, "nl-NL"]
-            ],
-        ],
-    )
-    async def test_generate(
-        self,
-        expected: str,
-        resource: str,
-        media_type: MediaType,
-        absolute: bool,
-        locale: ResolvableLocale | None,
-        additional_project_locale: Locale | None,
-        fragment: str | None,
-        query: Mapping[str, Sequence[str]] | None,
-        isolated_project_factory: IsolatedProjectFactory,
-    ) -> None:
-        locales = [default_locale]
-        if additional_project_locale:
-            locales.append(additional_project_locale)
-        async with isolated_project_factory(locales=locales) as project:
-            sut = await _StaticPathUrlUrlGenerator.new(project)
-            assert (
-                sut.generate(
-                    resource,
-                    absolute=absolute,
-                    fragment=fragment,
-                    locale=locale,
-                    media_type=media_type,
-                    query=query,
-                )
-                == expected
-            )
-
-
-@pytest.mark.parametrize(
-    ("expected", "resource"),
-    [
-        (True, DummyEntityOne()),
-        (True, "betty://some/path/index.html"),
-        (True, "betty:///some/path/index.html"),
-        (True, "betty-static://some/path/index.html"),
-        (True, "betty-static:///some/path/index.html"),
-        (False, ""),
-        (False, "/"),
-        (False, "index.html"),
-        (False, "example"),
-        (False, "/example"),
-        (False, "example/"),
-        (False, "/example/"),
-        (False, "example/index.html"),
-        (False, "/example/index.html"),
-        (False, object()),
-    ],
-)
-async def test_new_project_url_generator__supports(
-    expected: bool, resource: Any, isolated_project_factory: IsolatedProjectFactory
-) -> None:
-    async with isolated_project_factory(
-        plugins={EntityDefinition: [DummyEntityOne]}
-    ) as project:
-        sut = await new_project_url_generator(project)
-        assert sut.supports(resource) == expected
-
-
-@pytest.mark.parametrize(
-    (
-        "expected",
-        "clean_urls",
-        "resource",
-        "media_type",
-        "absolute",
-        "locale",
-        "additional_project_locale",
-    ),
-    [
-        # Entities
-        (
-            "https://example.com/dummy-one/my-first-entity/index.html",
-            False,
-            DummyEntityOne(id="my-first-entity"),
-            HTML,
-            True,
-            None,
-            None,
-        ),
-        # betty:// URLs
-        (
-            "https://example.com/some/path/index.html",
-            False,
-            "betty:///some/path/index.html",
-            HTML,
-            True,
-            None,
-            None,
-        ),
-        # betty-static:// URLs
-        (
-            "https://example.com/some/path/index.html",
-            False,
-            "betty-static:///some/path/index.html",
-            HTML,
-            True,
-            None,
-            None,
-        ),
-    ],
-)
-async def test_new_project_url_generator__generate(
-    expected: str,
-    clean_urls: bool,
-    resource: str,
-    media_type: MediaType,
-    absolute: bool,
-    locale: ResolvableLocale | None,
-    additional_project_locale: Locale | None,
-    isolated_project_factory: IsolatedProjectFactory,
-) -> None:
-    locales = [default_locale]
-    if additional_project_locale:
-        locales.append(additional_project_locale)
-    async with isolated_project_factory(
-        clean_urls=clean_urls,
-        locales=locales,
-        plugins={EntityDefinition: [DummyEntityOne]},
-    ) as project:
-        sut = await new_project_url_generator(project)
-        assert (
-            sut.generate(
-                resource,
-                media_type=media_type,
-                absolute=absolute,
-                locale=locale,
-            )
-            == expected
-        )
