+++ betty/tests/renderers/test_html.py
@@ -1,21 +1,22 @@
 from collections.abc import Mapping, Sequence
-from typing import Any, override
+from typing import Any, TypeGuard, override
 
 from betty.locale import ResolvableLocale
 from betty.media_type import ResolvableMediaType
 from betty.renderers.html import Html
-from betty.url import UrlGenerator
+from betty.url_generator import UrlGenerator
 
 
-class _TestHtmlUrlGenerator(UrlGenerator):
+class _TestHtmlUrlGenerator(UrlGenerator[str]):
     @override
-    def supports(self, resource: Any, /) -> bool:
+    def supports(self, resource: Any, /) -> TypeGuard[str]:
         return isinstance(resource, str) and resource == "betty-test://"
 
     @override
     def generate(
         self,
         resource: str,
+        /,
         *,
         absolute: bool = False,
         fragment: str | None = None,
