+++ betty/url_generators/__init__.py
@@ -1,0 +1,3 @@
+"""
+Reusable URL generators.
+"""
