+++ documentation/usage/url.rst
@@ -14,7 +14,7 @@ How to generate URLs
 Python
 ^^^^^^
 
-URLs are generated by calling the :py:meth:`betty.url.UrlGenerator.generate` method of any URL generator. In most cases,
+URLs are generated by calling the :py:meth:`betty.url_generator.UrlGenerator.generate` method of any URL generator. In most cases,
 you will want to use :py:attr:`betty.project.Project.url_generator`, which supports all project resources.
 
 .. code-block:: python
@@ -30,7 +30,7 @@ Jinja2 templates
     resource | url
 
 The :py:func:`url <betty.jinja_filters.url.Url>` filter works almost identical to
-:py:meth:`betty.url.UrlGenerator.generate`, except that it fills in some defaults from the Jinja2 environment, such as
+:py:meth:`betty.url_generator.UrlGenerator.generate`, except that it fills in some defaults from the Jinja2 environment, such as
 the template's current locale and media type. This means that if you are rendering an HTML page, and want to link to
 another HTML page in the same language, you will not need to provide this information explicitly.
